Add attendance-status.enum.ts file to enums folder
