+++ src/enums/attendance-status.enum.ts
@@ -1,0 +1,7 @@
+export enum AttendanceStatus {
+  PRESENT = 'present',
+  ABSENT = 'absent',
+  LATE = 'late',
+  EXCUSED = 'excused',
+  PENDING = 'pending',
+}
